requirements.txt
flask
google-cloud-storage
pillow
pandas
requests